Fix show event information when not enough space in calendar
Calendar has already an absolute position and z-index set so when a child element (pippy elment) also sets a new z-index that won't work due to how stacking works in browsers

What I did is hook the pipppy instances to an element outside of the calendar events area. That way it wont be affected by the previuos set position / z-index values

Also had to give a higher z-index to pippy instances because the full calendar popover had the same z-index as pippy and the browser stacking process would give higher priority to the full calendar popover because it was after the pippy element.