Added `bg:` prefix to events
Adding the `bg:` prefix to an event's `StateField` field allows you to force the background of the event's day to be of that color. If two events with the `bg:` prefix fall on the same day, the last event wins (the last event is the last on the listing of events, in the dashboard instance). **The background is not applied on weekends, purposefully. This is because this feature was initially inserted to represent holidays.**