Fixed 'null pointer' in Kibana component
When the kibana component was taking the longest to load and the viewport varies, we setup a series of timeouts. In some of these timeouts - for some reason - the window in the iframe was not yet loaded. By adding a verification, we simply do nothing in these situations and hope future timeout handles it.